:recycle: Refactor #25: LabelEditModal 공통 컴포넌트로 분리
