+++ app/src/main/java/com/umc/edison/presentation/label/LabelEditMode.kt
@@ -1,0 +1,5 @@
+package com.umc.edison.presentation.label
+
+enum class LabelEditMode {
+    NONE, ADD, EDIT, DELETE
+}
